+++ src/settings/MappingProfiles/initialDetails/MARC_BIBLIOGRAPHIC.js
@@ -1,7 +1,10 @@
 const MARC_BIBLIOGRAPHIC = {
   name: 'marcBib',
   recordType: 'MARC_BIBLIOGRAPHIC',
-  marcMappingDetails: [{ order: 0 }],
+  marcMappingDetails: [{
+    order: 0,
+    field: { subfields: [{}] },
+  }],
 };
 
 export default MARC_BIBLIOGRAPHIC;